requirements.txt
fastmcp>=2.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
faster-whisper>=0.9.0
openpyxl>=3.1.0
python-docx>=0.8.0
pdfplumber>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
GitPython>=3.1.0
schedule>=1.2.0
flask>=3.0.0
PyYAML>=6.0.0
tqdm>=4.66.0
pytest>=8.0.0
uvicorn>=0.30.0
Pillow>=10.0.0
pytesseract>=0.3.10
